+++ recipes/configs/dev/llama3/8B_dora_fsdp2.yaml
@@ -61,7 +61,7 @@ loss:
 
 # Training
 epochs: 1
-max_steps_per_epoch: 1
+max_steps_per_epoch: null
 gradient_accumulation_steps: 1
 
 # Logging
