+++ torchtune/modules/peft/__init__.py
@@ -17,7 +17,7 @@ from ._utils import (  # noqa
     validate_state_dict_for_lora,
 )
 from .dora import DoRALinear
-from .lora import _lora_a_init_params, _lora_b_init_params, LoRALinear
+from .lora import LoRALinear
 
 
 __all__ = [
@@ -32,6 +32,4 @@ __all__ = [
     "disable_adapter",
     "get_merged_lora_ckpt",
     "get_lora_module_names",
-    "_lora_a_init_params",
-    "_lora_b_init_params",
 ]
